requirements.txt
python-usbtmc
matplotlib
pyusb
libusb
numpy
#pyvisa-py
#pandas
scipy